+++ frontend/package.json
@@ -19,8 +19,7 @@
     "axios": "1.7.2",
     "react": "18.3.1",
     "react-dom": "18.3.1",
-    "react-router-dom": "6.23.1",
-    "styled-components": "6.1.11"
+    "react-router-dom": "6.23.1"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "6.4.6",
