+++ frontend/src/assets/Leftovers Website Logo.png
